client/requirements.txt
selenium==4.16.0
requests==2.31.0
python-socketio[client]==5.10.0
pywin32==306
